+++ server/stockapi.php
@@ -12,7 +12,7 @@ if (isset($_POST['stockId'])) {
         for($i=0;$i<mysqli_num_rows($data);$i++){
             $dump = mysqli_fetch_assoc($data);
             foreach ($dump as $key => $value) {
-                $dump[$key] = (int)$value;
+                $dump[$key] = (float)$value;
             }
             $load[$i] = $dump;
         }
